+++ lua/terrortown/menus/gamemode/administration/mapentities.lua
@@ -2,7 +2,7 @@
 
 CLGAMEMODESUBMENU.base = "base_gamemodesubmenu"
 
-CLGAMEMODESUBMENU.priority = 91
+CLGAMEMODESUBMENU.priority = 90
 CLGAMEMODESUBMENU.title = "submenu_administration_mapentities_title"
 
 function CLGAMEMODESUBMENU:Populate(parent)