+++ lua/terrortown/menus/gamemode/administration/roundsetup.lua
@@ -6,69 +6,51 @@ CLGAMEMODESUBMENU.priority = 98
 CLGAMEMODESUBMENU.title = "submenu_administration_round_setup_title"
 
 function CLGAMEMODESUBMENU:Populate(parent)
-	local form = vgui.CreateTTT2Form(parent, "header_round_setup_plyspawn")
+	local form = vgui.CreateTTT2Form(parent, "header_round_dead_players")
 
-	form:MakeHelp({
-		label = "help_ply_spawn"
-	})
-
-	form:MakeSlider({
-		serverConvar = "ttt_armor_on_spawn",
-		label = "label_armor_on_spawn",
-		min = 0,
-		max = 250,
-		decimal = 0
-	})
-
-	form:MakeHelp({
-		label = "help_hint_armor"
-	})
-
-	local form2 = vgui.CreateTTT2Form(parent, "header_round_dead_players")
-
-	form2:MakeCheckBox({
+	form:MakeCheckBox({
 		serverConvar = "ttt_identify_body_woconfirm",
 		label = "label_identify_body_woconfirm"
 	})
 
-	form2:MakeCheckBox({
+	form:MakeCheckBox({
 		serverConvar = "ttt_announce_body_found",
 		label = "label_announce_body_found"
 	})
 
-	form2:MakeCheckBox({
+	form:MakeCheckBox({
 		serverConvar = "ttt2_confirm_killlist",
 		label = "label_confirm_killlist"
 	})
 
-	form2:MakeCheckBox({
+	form:MakeCheckBox({
 		serverConvar = "ttt2_inspect_detective_only",
 		label = "label_inspect_detective_only"
 	})
 
-	form2:MakeCheckBox({
+	form:MakeCheckBox({
 		serverConvar = "ttt2_confirm_detective_only",
 		label = "label_confirm_detective_only"
 	})
 
-	form2:MakeCheckBox({
+	form:MakeCheckBox({
 		serverConvar = "ttt_dyingshot",
 		label = "label_dyingshot"
 	})
 
-	local form3 = vgui.CreateTTT2Form(parent, "header_round_setup_prep")
+	local form2 = vgui.CreateTTT2Form(parent, "header_round_setup_prep")
 
-	form3:MakeCheckBox({
+	form2:MakeCheckBox({
 		serverConvar = "ttt2_prep_respawn",
 		label = "label_prep_respawn"
 	})
 
-	form3:MakeCheckBox({
+	form2:MakeCheckBox({
 		serverConvar = "ttt_nade_throw_during_prep",
 		label = "label_nade_throw_during_prep"
 	})
 
-	form3:MakeSlider({
+	form2:MakeSlider({
 		serverConvar = "ttt_preptime_seconds",
 		label = "label_preptime_seconds",
 		min = 0,
@@ -76,7 +58,7 @@ function CLGAMEMODESUBMENU:Populate(parent)
 		decimal = 0
 	})
 
-	form3:MakeSlider({
+	form2:MakeSlider({
 		serverConvar = "ttt_firstpreptime",
 		label = "label_firstpreptime_seconds",
 		min = 0,
@@ -84,9 +66,9 @@ function CLGAMEMODESUBMENU:Populate(parent)
 		decimal = 0
 	})
 
-	local form4 = vgui.CreateTTT2Form(parent, "header_round_setup_round")
+	local form3 = vgui.CreateTTT2Form(parent, "header_round_setup_round")
 
-	form4:MakeSlider({
+	form3:MakeSlider({
 		serverConvar = "ttt_roundtime_minutes",
 		label = "label_roundtime_minutes",
 		min = 0,
@@ -94,16 +76,16 @@ function CLGAMEMODESUBMENU:Populate(parent)
 		decimal = 0,
 	})
 
-	form4:MakeHelp({
+	form3:MakeHelp({
 		label = "help_haste_mode"
 	})
 
-	local enbHaste = form4:MakeCheckBox({
+	local enbHaste = form3:MakeCheckBox({
 		serverConvar = "ttt_haste",
 		label = "label_haste"
 	})
 
-	form4:MakeSlider({
+	form3:MakeSlider({
 		serverConvar = "ttt_haste_starting_minutes",
 		label = "label_haste_starting_minutes",
 		min = 0,
@@ -112,7 +94,7 @@ function CLGAMEMODESUBMENU:Populate(parent)
 		master = enbHaste
 	})
 
-	form4:MakeSlider({
+	form3:MakeSlider({
 		serverConvar = "ttt_haste_minutes_per_death",
 		label = "label_haste_minutes_per_death",
 		min = 0,
@@ -121,16 +103,16 @@ function CLGAMEMODESUBMENU:Populate(parent)
 		master = enbHaste
 	})
 
-	form4:MakeHelp({
+	form3:MakeHelp({
 		label = "help_sherlock_mode"
 	})
 
-	form4:MakeCheckBox({
+	form3:MakeCheckBox({
 		serverConvar = "ttt_sherlock_mode",
 		label = "label_sherlock_mode"
 	})
 
-	form4:MakeSlider({
+	form3:MakeSlider({
 		serverConvar = "ttt_minimum_players",
 		label = "label_minimum_players",
 		min = 0,
@@ -138,14 +120,14 @@ function CLGAMEMODESUBMENU:Populate(parent)
 		decimal = 0
 	})
 
-	local form5 = vgui.CreateTTT2Form(parent, "header_round_setup_post")
+	local form4 = vgui.CreateTTT2Form(parent, "header_round_setup_post")
 
-	form5:MakeCheckBox({
+	form4:MakeCheckBox({
 		serverConvar = "ttt_postround_dm",
 		label = "label_postround_dm"
 	})
 
-	form5:MakeSlider({
+	form4:MakeSlider({
 		serverConvar = "ttt_posttime_seconds",
 		label = "label_posttime_seconds",
 		min = 0,
@@ -153,13 +135,13 @@ function CLGAMEMODESUBMENU:Populate(parent)
 		decimal = 0
 	})
 
-	local form6 = vgui.CreateTTT2Form(parent, "header_round_setup_map_duration")
+	local form5 = vgui.CreateTTT2Form(parent, "header_round_setup_map_duration")
 
-	form6:MakeHelp({
+	form5:MakeHelp({
 		label = "help_round_limit"
 	})
 
-	form6:MakeSlider({
+	form5:MakeSlider({
 		serverConvar = "ttt_round_limit",
 		label = "label_round_limit",
 		min = 0,
@@ -167,7 +149,7 @@ function CLGAMEMODESUBMENU:Populate(parent)
 		decimal = 0
 	})
 
-	form6:MakeSlider({
+	form5:MakeSlider({
 		serverConvar = "ttt_time_limit_minutes",
 		label = "label_time_limit_minutes",
 		min = 0,